backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pymongo>=4.8.0,<5.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.9,<1.0.0
python-dotenv>=1.0.1,<2.0.0
cloudinary>=1.40.0,<2.0.0
email-validator>=2.2.0,<3.0.0
